% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_group_means.R
\name{calculate_group_means}
\alias{calculate_group_means}
\title{Creating a Function to Calculate Means by Group}
\usage{
calculate_group_means(data, group_name, quantity_name)
}
\arguments{
\item{data}{Tibble data frame that contains the grouping and the desired quantity to calculate.}

\item{group_name}{name of the column in data that contains the grouping, it does not have to be a string.}

\item{quantity_name}{name of the column in data that contains the quantities for mean calculations}
}
\value{
An object of class "Tibble" that returns
the means of each group.
}
\description{
Allow for an easier way to calculate the means
grouped by a certain categorical variable. It also prints a boxplots of the grouped
quantitative values.
}
\examples{
calculate_group_means(InsectSprays, spray, count)
}
